requirements_only_minimal.txt
requests
python-dotenv
apify-client==2.0.1
urllib3>=2.0.0